+++ src/check-dependencies-defaults.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const path = require('path');
+const processCwd = process.cwd();
+
+module.exports = {
+  files: require('./lib/global-source-files'),
+  packageFile: path.join(processCwd, 'package.json')
+};